chahua: 重构 renderer.js（五）—— 抽出 envelope_router 模块
ws 下行 envelope 的总分派（~200 行 handleEnvelope + formatGuestCaps）抽到
envelope_router.js。foregroundRoomId（P9 前台房 room_id，后台里程碑分流基准）
全封进模块；ROOM_EXPORT / FILE_DOWNLOAD 的 Blob 下载收敛成 triggerDownload 一个
helper。turn_start/turn_end 的打分写经 applyScores 回调（scoresByName 仍归
renderer），其余按 type 分发到各 feature 模块经 deps 注入。

taskState / handoffState 是 ES module 单例 —— envelope_router 直接 import，与
renderer 观察的是同一份状态。renderer 内 6 个随之失用的 events.js 导入清掉。

renderer.js 1309 → 1110 行（自原始 1671 累计 −561，−34%）。行为零变化，
node --check 通过。
